separate optimized Array classes and Query classes
